Added: - Some new theme color settings - Firebase database connection - Login and register pages
There have also been some adjustments to the way users get stored and read in the firebase database. Users can log in using their email address so I've made the path to the relevant user data a random id instead of the username or email address. This prevents users encountering errors when multiple users use the same username and/or email address.